Correct PAN to +100 (100% R) on AUX 6 (Bkg Music R)
Today the PAN on AUX 6 (Bkg Music R) was at -100 (full left)
Hence, there were no music coming out the right speaker.
Since the mixing table is on the left, it was only noticeable by
looking at the meter.

Note that was edited manually since using a M32 Edit for linux
result in a lot of changes.