Optimize livestream mix for audience microphones (v.7.1.0) (#26)
# Changes

## Optimize the leveling in the livestream

Changes in the audience will now more likely have a great impact to the livestream as well. It still makes sense to check the balances in the livestream during sound check. Best case is always to have someone taking care of the mix in the livestream.

### Effect levels in the livestream

The effects busses are now completely separated.

#### Room FX

The drums (i.e., snare, toms, bit of overheads), the keys, and the acoustic guitar use this effect. We removed the vocals from this bus.

#### Voc FX

Only the vocals go into this bus.

### Audience microphones

We ran the 2 new Røde shot gun mics (NTG-4) for the 1st time. The setup guide for the Stage Engineer (you can find it in our [LinkTree](https://linktr.ee/mosaikberlinsound)) explains how to set them up on the stands for the TVs and the [patch field](
![](https://github.com/mosaikberlinsound/m32-scenes-and-presets/blob/main/patch-field/Patch%20Field.png)
) is updated accordingly. We connect the audience microphones to channel 28/29. On the sound board these are mapped to channels 29 and 30. Please be aware that the new DCA group “Audience” needs to stay at around -5 dB to be able to hear the audience in the livestream. We set it up as post fader in order to have quick control over the audience volume in the live stream.

### Bus/Matrix levels and their compressors and limiters

We adjusted the compressor and limiter so that the volume difference between worship and preach is not too dramatic. The Bus 11/12 has a compressor with a high ratio of 7:1 and a threshold of -33 dB. The matrix functions as a limiter with a ratio of 5:1 at a threshold of -10 dB.

I created #27 to further check if this orchestration needs to be optimized.

## Sacrificing Delay FX

The DCA group 8 was used to control the delay FX send. I rarely see someone using the delay FX so I thought it is fine to sacrifice the DCA. Accordingly, the bus send for FX 16 is off now, so we do not accidentally have a delay FX with the wrong BPM.

# Fixes

- #24 
- #25