Introduction of new drum kit and kick mic (#31)
## New drum kit

Our new drum kit which we purchased earlier this year needs different settings for EQ, compression etc. as the drum kit we had before.

## New kick mic

We introduced the Shure Beta 91A as an additional kick mic. The AUDIX mic stays as our mic to get body of the kick drum. What was missing was the kick itself. It wasn't very precise. Therefore, we introduce the Shure mic as an addition. Its only purpose is to get a more precise kick sound. This change required modifications on the patching and therefore we have a new patch field sheet which shows which microphone or instrument is connected at which place.