Changed reverb effects for vocals and instruments
We selected different effects for vocals and instruments and adjusted them so the instruments and the vocals sound better in the livestream. The effects usually have minimal impact on the room sound but their effect is very useful for the livestream.
For the instruments we use the Ambience effect (Bus 14 - ROOM REV).
For the vocals we use the Rich Plate Reverb (Bus 15 - VOX REV).